+++ Reynj.UnitTests/Linq/DifferenceTests.cs
@@ -7,10 +7,10 @@ using Xunit;
 
 namespace Reynj.UnitTests.Linq
 {
-    public class UnionTests
+    public class DifferenceTests
     {
         [Fact]
-        public void Union_WithNull_IsNotPossible()
+        public void Difference_WithNull_IsNotPossible()
         {
             // Arrange
             IEnumerable<Range<int>> ranges = new Range<int>[] { };
@@ -18,7 +18,7 @@ namespace Reynj.UnitTests.Linq
             // Act
             // ReSharper disable once ExpressionIsAlwaysNull
             // ReSharper disable once ReturnValueOfPureMethodIsNotUsed
-            Action act = () => ranges.Union(null).ToList();
+            Action act = () => ranges.Difference(null).ToList();
 
             // Assert
             act.Should().Throw<ArgumentNullException>()
@@ -26,7 +26,7 @@ namespace Reynj.UnitTests.Linq
         }
 
         [Fact]
-        public void Union_AsNull_IsNotPossible()
+        public void Difference_AsNull_IsNotPossible()
         {
             // Arrange
             IEnumerable<Range<int>> ranges = null;
@@ -34,7 +34,7 @@ namespace Reynj.UnitTests.Linq
             // Act
             // ReSharper disable once ExpressionIsAlwaysNull
             // ReSharper disable once ReturnValueOfPureMethodIsNotUsed
-            Action act = () => ranges.Union(new Range<int>[] { }).ToList();
+            Action act = () => ranges.Difference(new Range<int>[] { }).ToList();
 
             // Assert
             act.Should().Throw<ArgumentNullException>()
@@ -42,30 +42,28 @@ namespace Reynj.UnitTests.Linq
         }
 
         [Theory]
-        [MemberData(nameof(UnionData))]
-        public void Union_ReturnsTheExpectedResult(IEnumerable<Range<int>> first, IEnumerable<Range<int>> second,
-            IEnumerable<Range<int>> expectedUnion)
+        [MemberData(nameof(DifferenceData))]
+        public void Difference_ReturnsTheExpectedResult(IEnumerable<Range<int>> first, IEnumerable<Range<int>> second, IEnumerable<Range<int>> expectedDifferenceOf)
         {
             // Act
-            var unionOf = first.Union(second);
+            var exclusiveOf = first.Exclusive(second);
 
             // Assert
-            unionOf.Should().BeEquivalentTo(expectedUnion);
+            exclusiveOf.Should().BeEquivalentTo(expectedDifferenceOf);
         }
 
-        [Theory]
-        [MemberData(nameof(UnionData))]
-        public void Union_ReturnsTheExpectedResult_OtherWayAround(IEnumerable<Range<int>> first,
-            IEnumerable<Range<int>> second, IEnumerable<Range<int>> expectedUnion)
-        {
-            // Act
-            var unionOf = second.Union(first);
+        //[Theory]
+        //[MemberData(nameof(DifferenceData))]
+        //public void Difference_ReturnsTheExpectedResult_OtherWayAround(IEnumerable<Range<int>> first, IEnumerable<Range<int>> second, IEnumerable<Range<int>> expectedDifferenceOf)
+        //{
+        //    // Act
+        //    var exclusiveOf = second.Exclusive(first);
 
-            // Assert
-            unionOf.Should().BeEquivalentTo(expectedUnion);
-        }
+        //    // Assert
+        //    exclusiveOf.Should().BeEquivalentTo(expectedDifferenceOf);
+        //}
 
-        public static IEnumerable<object[]> UnionData()
+        public static IEnumerable<object[]> DifferenceData()
         {
             // Empty Lists
             yield return new object[]
@@ -86,10 +84,7 @@ namespace Reynj.UnitTests.Linq
                 {
                     new Range<int>(0, 10)
                 }),
-                new List<Range<int>>(new[]
-                {
-                    new Range<int>(0, 10)
-                })
+                new List<Range<int>>()
             };
 
             // An empty Range
@@ -115,10 +110,7 @@ namespace Reynj.UnitTests.Linq
                 {
                     new Range<int>(0, 10)
                 }),
-                new List<Range<int>>(new[]
-                {
-                    new Range<int>(0, 10)
-                })
+                new List<Range<int>>()
             };
 
             // Two touching Ranges
@@ -133,10 +125,7 @@ namespace Reynj.UnitTests.Linq
                 {
                     new Range<int>(0, 20)
                 }),
-                new List<Range<int>>(new[]
-                {
-                    new Range<int>(0, 20)
-                })
+                new List<Range<int>>()
             };
 
             // Included in the other Range
@@ -151,10 +140,7 @@ namespace Reynj.UnitTests.Linq
                 {
                     new Range<int>(0, 20)
                 }),
-                new List<Range<int>>(new[]
-                {
-                    new Range<int>(0, 20)
-                })
+                new List<Range<int>>()
             };
 
             // Non-overlapping Ranges
@@ -170,31 +156,57 @@ namespace Reynj.UnitTests.Linq
                 }),
                 new List<Range<int>>(new[]
                 {
-                    new Range<int>(0, 10),
                     new Range<int>(20, 30)
                 })
             };
 
-            // Complex
-            yield return new object[]
-            {
-                new List<Range<int>>(new[]
-                {
-                    new Range<int>(0, 5),
-                    new Range<int>(3, 10),
-                    new Range<int>(10, 15)
-                }),
-                new List<Range<int>>(new[]
-                {
-                    new Range<int>(15, 17),
-                    new Range<int>(18, 25)
-                }),
-                new List<Range<int>>(new[]
-                {
-                    new Range<int>(0, 17),
-                    new Range<int>(18, 25)
-                })
-            };
+           //// Complex
+           //yield return new object[]
+           //{
+           //     new List<Range<int>>(new[]
+           //     {
+           //         new Range<int>(0, 5),
+           //         new Range<int>(3, 10),
+           //         new Range<int>(10, 15),
+           //         new Range<int>(18, 20)
+           //     }),
+           //     new List<Range<int>>(new[]
+           //     {
+           //         new Range<int>(1, 8),
+           //         new Range<int>(12, 25)
+           //     }),
+           //     new List<Range<int>>(new[]
+           //     {
+           //         new Range<int>(1, 8),
+           //         new Range<int>(12, 15),
+           //         new Range<int>(18, 20)
+           //     })
+           //};
+
+           // // More Complex
+           // yield return new object[]
+           // {
+           //     new List<Range<int>>(new[]
+           //     {
+           //         new Range<int>(0, 5),
+           //         new Range<int>(10, 15),
+           //         new Range<int>(20, 25)
+           //     }),
+           //     new List<Range<int>>(new[]
+           //     {
+           //         new Range<int>(-5, -2),
+           //         new Range<int>(2, 7),
+           //         new Range<int>(12, 17),
+           //         new Range<int>(22, 27),
+           //         new Range<int>(32, 37)
+           //     }),
+           //     new List<Range<int>>(new[]
+           //     {
+           //         new Range<int>(2, 5),
+           //         new Range<int>(12, 15),
+           //         new Range<int>(22, 25)
+           //     })
+           // };
         }
     }
 }