+++ Reynj.UnitTests/Linq/ExclusiveTests.cs
@@ -7,10 +7,10 @@ using Xunit;
 
 namespace Reynj.UnitTests.Linq
 {
-    public class UnionTests
+    public class ExclusiveTests
     {
         [Fact]
-        public void Union_WithNull_IsNotPossible()
+        public void Exclusive_WithNull_IsNotPossible()
         {
             // Arrange
             IEnumerable<Range<int>> ranges = new Range<int>[] { };
@@ -18,7 +18,7 @@ namespace Reynj.UnitTests.Linq
             // Act
             // ReSharper disable once ExpressionIsAlwaysNull
             // ReSharper disable once ReturnValueOfPureMethodIsNotUsed
-            Action act = () => ranges.Union(null).ToList();
+            Action act = () => ranges.Exclusive(null).ToList();
 
             // Assert
             act.Should().Throw<ArgumentNullException>()
@@ -26,7 +26,7 @@ namespace Reynj.UnitTests.Linq
         }
 
         [Fact]
-        public void Union_AsNull_IsNotPossible()
+        public void Exclusive_AsNull_IsNotPossible()
         {
             // Arrange
             IEnumerable<Range<int>> ranges = null;
@@ -34,7 +34,7 @@ namespace Reynj.UnitTests.Linq
             // Act
             // ReSharper disable once ExpressionIsAlwaysNull
             // ReSharper disable once ReturnValueOfPureMethodIsNotUsed
-            Action act = () => ranges.Union(new Range<int>[] { }).ToList();
+            Action act = () => ranges.Exclusive(new Range<int>[] { }).ToList();
 
             // Assert
             act.Should().Throw<ArgumentNullException>()
@@ -42,30 +42,28 @@ namespace Reynj.UnitTests.Linq
         }
 
         [Theory]
-        [MemberData(nameof(UnionData))]
-        public void Union_ReturnsTheExpectedResult(IEnumerable<Range<int>> first, IEnumerable<Range<int>> second,
-            IEnumerable<Range<int>> expectedUnion)
+        [MemberData(nameof(ExclusiveData))]
+        public void Exclusive_ReturnsTheExpectedResult(IEnumerable<Range<int>> first, IEnumerable<Range<int>> second, IEnumerable<Range<int>> expectedExclusiveOf)
         {
             // Act
-            var unionOf = first.Union(second);
+            var exclusiveOf = first.Exclusive(second);
 
             // Assert
-            unionOf.Should().BeEquivalentTo(expectedUnion);
+            exclusiveOf.Should().BeEquivalentTo(expectedExclusiveOf);
         }
 
         [Theory]
-        [MemberData(nameof(UnionData))]
-        public void Union_ReturnsTheExpectedResult_OtherWayAround(IEnumerable<Range<int>> first,
-            IEnumerable<Range<int>> second, IEnumerable<Range<int>> expectedUnion)
+        [MemberData(nameof(ExclusiveData))]
+        public void Exclusive_ReturnsTheExpectedResult_OtherWayAround(IEnumerable<Range<int>> first, IEnumerable<Range<int>> second, IEnumerable<Range<int>> expectedExclusiveOf)
         {
             // Act
-            var unionOf = second.Union(first);
+            var exclusiveOf = second.Exclusive(first);
 
             // Assert
-            unionOf.Should().BeEquivalentTo(expectedUnion);
+            exclusiveOf.Should().BeEquivalentTo(expectedExclusiveOf);
         }
 
-        public static IEnumerable<object[]> UnionData()
+        public static IEnumerable<object[]> ExclusiveData()
         {
             // Empty Lists
             yield return new object[]
@@ -168,31 +166,54 @@ namespace Reynj.UnitTests.Linq
                 {
                     new Range<int>(20, 30)
                 }),
-                new List<Range<int>>(new[]
-                {
-                    new Range<int>(0, 10),
-                    new Range<int>(20, 30)
-                })
+                new List<Range<int>>()
             };
 
-            // Complex
+           // Complex
+           yield return new object[]
+           {
+                new List<Range<int>>(new[]
+                {
+                    new Range<int>(0, 5),
+                    new Range<int>(3, 10),
+                    new Range<int>(10, 15),
+                    new Range<int>(18, 20)
+                }),
+                new List<Range<int>>(new[]
+                {
+                    new Range<int>(1, 8),
+                    new Range<int>(12, 25)
+                }),
+                new List<Range<int>>(new[]
+                {
+                    new Range<int>(1, 8),
+                    new Range<int>(12, 15),
+                    new Range<int>(18, 20)
+                })
+           };
+
+            // More Complex
             yield return new object[]
             {
                 new List<Range<int>>(new[]
                 {
                     new Range<int>(0, 5),
-                    new Range<int>(3, 10),
-                    new Range<int>(10, 15)
+                    new Range<int>(10, 15),
+                    new Range<int>(20, 25)
                 }),
                 new List<Range<int>>(new[]
                 {
-                    new Range<int>(15, 17),
-                    new Range<int>(18, 25)
+                    new Range<int>(-5, -2),
+                    new Range<int>(2, 7),
+                    new Range<int>(12, 17),
+                    new Range<int>(22, 27),
+                    new Range<int>(32, 37)
                 }),
                 new List<Range<int>>(new[]
                 {
-                    new Range<int>(0, 17),
-                    new Range<int>(18, 25)
+                    new Range<int>(2, 5),
+                    new Range<int>(12, 15),
+                    new Range<int>(22, 25)
                 })
             };
         }