+++ Reynj/Linq/Difference.cs
@@ -10,15 +10,16 @@ namespace Reynj.Linq
     public static partial class Enumerable
     {
         /// <summary>
-        /// Unions two sequences, meanwhile also reducing the result (AND)
+        /// Keeps the elements which are in either of the sequences and not in their intersection. (XOR)
         /// </summary>
-        /// <see href="https://en.wikipedia.org/wiki/Union_(set_theory)"/>
-        /// <param name="first">The first sequence to union.</param>
-        /// <param name="second">The sequence to union to the first sequence.</param>
+        /// <see href="https://en.wikipedia.org/wiki/Symmetric_difference"/>
+        /// <see href="https://en.wikipedia.org/wiki/Exclusive_or"/>
+        /// <param name="first">The first sequence.</param>
+        /// <param name="second">The sequence to do the exclusive comparison with.</param>
         /// <typeparam name="T">The type of the elements of the input sequences.</typeparam>
-        /// <returns>An <see cref="T:System.Collections.Generic.IEnumerable`1"></see> that contains the unionized elements of the two input sequences.</returns>
+        /// <returns>An <see cref="T:System.Collections.Generic.IEnumerable`1"></see> that contains the exclusive elements of the two input sequences.</returns>
         /// <exception cref="T:System.ArgumentNullException"><paramref name="first">first</paramref> or <paramref name="second">second</paramref> is null.</exception>
-        public static IEnumerable<Range<T>> Union<T>(this IEnumerable<Range<T>> first, IEnumerable<Range<T>> second)
+        public static IEnumerable<Range<T>> Difference<T>(this IEnumerable<Range<T>> first, IEnumerable<Range<T>> second)
             where T : IComparable
         {
             if (first == null) 
@@ -26,9 +27,11 @@ namespace Reynj.Linq
             if (second == null) 
                 throw new ArgumentNullException(nameof(second));
 
-            return first
-                .Concat(second) // Merge the two range collections (Concat is used instead of Union, because the Reduce removes the doubles)
-                .Reduce(); // Merge overlapping and touching ranges
+            // By reducing both sequences the set difference logic becomes easier
+            var firstReduced = first.Reduce().ToList();
+            var secondReduced = second.Reduce().ToList();
+
+
         }
     }
 }