+++ CHANGELOG.md
@@ -5,7 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+<!--
 ### Added
+- The Difference method on a IEnumerable<Range<T>> to get the relative complement of two sequences and also reduce the result
+- The Exclusive method on a IEnumerable<Range<T>> to get the symmetric difference of two sequences and also reduce the result
+-->
 
 ### Fixed
 - IsEmpty() now works with reference types